Move Text to kas_core::theme
This makes two variants of Text accessible
