+++ crates/kas-core/src/decorations.rs
@@ -8,8 +8,7 @@
 //! Note: due to definition in kas-core, some widgets must be duplicated.
 
 use crate::event::{CursorIcon, ResizeDirection};
-use crate::text::Text;
-use crate::theme::TextClass;
+use crate::theme::{Text, TextClass};
 use kas::prelude::*;
 use kas::theme::MarkStyle;
 use kas_macros::impl_scope;
