+++ crates/kas-core/src/event/cx/config.rs
@@ -10,8 +10,8 @@ use crate::event::{EventState, FocusSource};
 use crate::geom::{Rect, Size};
 use crate::layout::AlignPair;
 use crate::messages::Erased;
-use crate::text::{format::FormattableText, Text, TextApi};
-use crate::theme::{Feature, SizeCx, ThemeSize};
+use crate::text::{format::FormattableText, TextApi};
+use crate::theme::{Feature, SizeCx, Text, ThemeSize};
 use crate::{Id, Node};
 use cast::Cast;
 use std::fmt::Debug;
